Refactor TCG log instruction backend into multiple files.
Each trace format is moved to a separate file for implementation.
